feat(zupdater): add release channel to builds to specify where zupdater looks for updates
This adds a "channel" configuration to releases that limits what
releases the updater will consider when looking for new versions. This
allows us to configure a "2.55" release channel, and prevents the
from updating into 3.0 releases while still allowing it to grab new
releases from 2.55.

This is necessary long term as it allows 2.55 users getting easy bug fix
updates without moving them to 3.0, which should only be done
explicitly.