misc!: add 'def_pair' container type.
Works like std::pair, but when constructed, both or just the second value can be left out to default-construct them.
Also adds packfile support for this type and 'std::pair'

(...std::pair docs SAYS it can be default constructed in this way, but I ran into compile errors doing so. Something to do with explicit maybe?)