feat(zc): optimize zasm scripts
First thing when loading a quest the ZASM scripts will undergo
optimizations. Most optimization passes apply improvements similar
to that which have been implemented in the zscript compiler, but are
not present in older quest scripts. Others are novel and improve even
modern zscript compiler output.

Savings for a few quests:

- Stellar Seas - 163762 ops (5%)
- Yuurad       - 604668 ops (4%)
- Umbral Cloud - 105994 ops (8%)

The optimizations done currently are:

- Removing unreachable blocks of code
- Replace many PUSH/POP with equivalent single PUSHARGSR/POPARGS
- Reduce ops used for conditionals (COMPARER)

This is on by default, but can be disabled with
`[ZSCRIPT] optimize_zasm = 0` in `zc.cfg`.