fix(zc): prevent crash in save creation by resetting global vars
When a game is loaded, many global variables are initialized to a known
state. These variables were not reset when the game ends and returns
to the title screen. However, save creation happens to call code that
checks the active subscreen pointer, which was invalidated. Accessing
this could randomly crash or corrupt the program.

Now, these global variables are reset also just before loading the
title screen.