refactor(zc): remove ZASM debugger from menu
You may still activate this with `[CONSOLE] print_ZASM = 1` in `zc.cfg`,
but the option is no longer in the player menu. This will be removed
entirely soon - it's so slow it can lock up a computer, and not useful
enough to keep around.

We may introduce a proper debugger in the future for ZScript to serve
the purpose the ZASM debugger tried to.