fix: retain monitor resolution for fullscreen mode
Previously, the maximum monitor resolution would be used when using
fullscreen mode. That was bad because it messes up the window layout of
anything else on that monitor. Now we create fullscreen displays without
changing the resolution.

Also, no longer destroy and recreate the display to change
fullscreen/windowed mode. Switching between is much smoother now, and
the code got a bit simpler and safer.