fix(zc): show final game frame when opening/closing subscreen
For older quests utilizing none of the QRs related to "cool scrolling",
the subscreen open/close crawl would show a version of the game frame
that was missing a few things, like enemy or player sprites.

This should allow us to simplify the rendering (no need to render most
things twice to a secondary bitmap), and just looks better.