misc(zscript): add Region in anticipation of z3 scrolling
To enable scripts to work in 2.55 and in a future release with z3
scrolling, we are adding a non-functional `Region`:

- `Region->Width`: width in pixels (today, always 256)
- `Region->Height`: height in pixels (176)
- `Region->ScreenWidth`: width in screens (1)
- `Region->ScreenHeight`: height in screens (1)
- `Region->NumCombos`: number of combo positions (176)
- `Region->ID`: region id (0 - meaning just a normal 1x1 area)
- `Region->OriginScreen`: top-left screen in the region
- `Region->GetScreenForComboPos(int pos)`: given a combo position, returns the screen index associated with it
- `Region->TriggerSecrets(int screen_index)`: trigger secrets for given screen. The screen must be in the region

For example, if a script uses `Region->NumCombos` instead of `176` to
iterate all the combos in the current area, it will work in both regions
and non-regions. Additionally you should use `ComboAt(x, y)` rather than
anything else to translate between pixel coordinates and a combo
position.