fix(zc): actually stop mp3/ogg, and explicitly stop title music
The titlescreen music would restart in the file select screen if the
Sound dialog was opened then closed. This was because the volume was set
to 0 to fade out, but it was never actually stopped.

But also, due to a bug ogg and mp3 formats were really just "paused"
when told to stop, so setting the volume afterwards would result in
resuming music that was meant to be over.