feat(zq): export/import strings to .tsv text file
Allows for exporting strings to a spreadsheet, text editor, or other
program for modification, then importing them back to ZC.

The first row of the `.tsv` (tab separated values) file are the column
names, followed by an example row to designate where the first three
lines begin and end for each string (but you can add more lines than
that).

For best results in a spreadsheet, "Freeze" the first two rows and use a
monospace font for the "message" column.

Example:
https://docs.google.com/spreadsheets/d/197-NSvNQJ174Ri9GVlxYDX4oFp8w1SZbmOXuDohKZZk/edit?usp=sharing