refactor(zscript): standardize script object management
This keeps script-created objects in the same map, gives all objects
a unique id, and implements allocation and retrieval for any object type
using the same methods. This reduces some duplicated code, and should
help with implementation of a potential garbage collector feature.

Bitmap objects are a bit special, and still need to be refactored.