fix(zscript): derive language runtime version from compiled scripts
The script engine checks the version a script was compiled in to vary
some behavior. Previously this version was set to the section version
in the qst file format, but that is updated even if scripts were not
re-compiled. Now we look directly at the script metadata to set this
version. This avoids issues with saving a quest in a new version without
a recompile.