refactor(zscript): move internal symbols to new binding .zh files
Internal symbols were defined in `.cpp` files, but now they are defined
in `.zs` files with a new `internal` keyword and `@zasm` doc comments.
These files live in `bindings`, and are automatically included for every
compilation.

This should simplify the process of adding new internal symbols. It also
enables viewing documentation of any internal symbol in the VS Code
extension (w/ a hover tooltip), and all the documentation we have
for internal functions and variables has been inserted as comments in
the binding files.

This also deprecates functions that create an object of internal types,
like `Game->CreateBitmap(width, height)` - and replaces with normal
constructors - `new bitmap(width, height)`.