refactor(zscript): make all command strings match enum name
This normalizes all the string names for commands to be the same
as the enum name used in code, and rewrites `get_script_command` to
be an explicit mapping of this string representation and the enum value.
These not matching was confusing, and was a potential source of future
bugs.