feat(zscript): add LEGACY_ARRAYS option as more lenient approach to array syntax
This replaces the `OLD_ARRAY_TYPECASTING` option with `LEGACY_ARRAYS`.
It can be set to `on`, `off`, or `warn`, and defaults to `warn`. This
option supports usage of deprecated syntax for arrays. This will
continue to work:

```
int arr = {1, 2};
Trace(arr[0]);
```

* by default, this will warn that `int[] arr` is the preferred syntax
* legacy arrays syntax is currently not allowed for objects that must be
  free'd, such as custom classes or `bitmap` (but, for example,
  `lweapon` is fine)

Additionally but independent of this option, parameter types are cast
to/from arrays when resolving a function call. There was an existing
feature that selects the function that required the fewest type casts,
and this taps into that. For example, this code works:

```
void fnArray(int[] arr) {}
void fnLegacyArray(int arr) {}

...

int legacy_arr = {1};
fnArray(arr); // arr is cast to int[]
int arr = {1};
fnLegacyArray(arr); // arr is cast to int
```