feat(zq): add "Auto-Save" and "Auto-Restart" options to test dialog
The test dialog no longer prompts you about unsaved changes. Instead, a
new option in the dialog "Auto-Save" will save for you just before
launching test mode. This defaults to on, but can be disabled.

Additionally, the existing "Auto-Restart" option was moved from the
launcher to the test dialog, where it belongs.