fix(zscript): fix broken array literals when declared as array pointers
In ZScript, there are array declarations and array pointers. Array
pointers are used as array parameters, and declarations are (typically)
used to declare arrays (either with an initializer list, or as an empty
array).

Array declarations register their array to be deallocated at the end of
the scope. Array pointers can take an array, and won't deallocate it
when it goes out of scope. That's why pointers are used in parameters.

This syntax is easy to mix up, so one might declare an array using the
pointer syntax. The compiler would not complain, but it would actually
produce invalid code.

```
item items1[] = {item}; // array declaration syntax, worked
item[] items1 = {item}; // array pointer syntax, was broken
```

This is a recent-ish regression in the 3.0 branch, and did not happpen
in 2.55. This has now been resolved.

end changelog

Some key implementation changes to note:

- `BuildOpcodes::caseDataDecl` now only used `buildVariable`, even for
  arrays
- local arrays are now registered for dealloc via
  `registerLocalArrayDealloc`, and that usually gets called right after
  `OAllocateMemImmediate` is emitted. The exception is arrayLiteralFree
  and stringLiteralFree, since they dealloc their temporary values
  sooner, and in a different way
- buildArrayUninit now pushes its value as EXP1, to be used by
  buildVariable.