fix(zq): remove hidden combo selection mode toggled with spacebar (#1051)
Pressing space changed how combo selection worked in a way that often confused
users. There was also no visible indication of this, so it was extra bad. This
mode was perhaps useful for something in the combo selector dialog, but outside
that it was seemingly not useful at all - and is now removed.