feat: z3 scrolling regions (#1074)
Regions are a collection of screens that are stitched together to become
one playing area that the player can move around in, with the viewport
moving to track the player. All mechanics of ZC have been updated to
support that there can now be multiple active screens. Additionally,
there is a new DMap option to reclaim 160 more pixels in height for the
visible playing field (making the passive subscreen draw over it). There
is also scripting support for configuring regions and the viewport.

Here is a
[guide](https://docs.zquestclassic.com/tutorials/z3_user_guide) covering
what quest makers need to know about regions.