feat: show bottom 8 pixels
The bottom 8 pixels of the screen has always been hidden. This was
considered to be a universally bad thing, so they are now visible. A QR
rule is added to control this behavior, and by default all existing
quests have been opted into it, except for those known to have issues
related to script drawing.

Active subscreens are still just 168 pixels tall. When the bottom 8
pixels are show, active subscreens travel an extra 8 pixels when opened.

ZC Player has a setting to configure this too. By default, it defers to
the QR as defined in the quest, but users can override this.