feat: show bottom 8 pixels
The bottom 8 pixels of the screen has always been hidden. This was
considered to be a universally bad thing, so they are now visible. A
compat quest rule is added to control this behavior.

When the bottom 8 pixels are enabled, active subscreens are still just
168 pixels tall. Instead, active subscreens travel an extra 8 pixels
when opened.

ZC Player has a setting to configure this too. By default, it defers to
the QR as defined in the quest, but users can override this.

(cherry picked from commit 3bec0750efe2b17f3a3f7a4182a8a47a73ca2035)