refactor(zscript): print context alongside script runtime errors
When a script does something that results in an error, a message is
logged to the console that usually mentions the zscript source variable
or function where the problem occurred. This was usually right, but
sometimes would mention the wrong thing.

Some automation is added to more consistently print the correct variable
or function name when encountering an error. In addition, potentially
multiple bits of context may now print alongside error messages (one
or all of: function, variable, or the name of a specific parameter).