feat(zscript): persist all reachable objects to save file
`GlobalObject` is no longer necessary to persist a custom object to save
files. It is now enough to assign it to a global variable or array - as
long as it uses the custom object type (rather than `int` or `untyped`).

This includes all nested objects, and any object reachable from an
object that was globalized via `GlobalObject`.

https://discord.com/channels/876899628556091432/1377446615178281190