fix(zscript)!: handle object references for buggy old init timing
When the "Use Old Global Init and SaveLoad Timing" compat QR is on,
if a script that runs before the global init script (such as a generic
script) assigns a global object, the engine would get confused and
crash.

This was because the global init script is what determines what global
registers are object types, and assigning an object to a register before
its been declared to be an object type fails an assertion.