refactor(zscript): improve stack overflow error behavior
Previously, if a script used too many stack variables (typically by
doing lots of recursion), the engine would (sometimes) emit a warning,
"wrap around" the stack pointer, and continue executing the script.
However, that results in very buggy and random behavior, and is never
recoverable.

Now when the stack overflows, the script is stopped.