test!: remove ZC_PACKAGE_REPLAYS
The web build no longer packages test replays and quests in the build
directly for the replay test runner. Instead, it can now load files from
the local machine. Now, any replay or quest can be loaded in the web
build for local development (does not have to be inside the tests
folder).