fix(zscript): make owned objects live as long as the sprite owner
The `Own` functions on sprite objects (npc, ffc, itemsprite, weapons)
are documented as assigning the object's owner to the sprite. However,
they were actually being "owned" by the sprite's script, so in the case
where a script ends before a sprite is destroyed, the object lost its
owner a little earlier than expected.