fix(zscript): clear eweapon script when it ends
When a script ends, the scripting engine is supposed to clear the
object's script so that it cannot run again. But a bug prevented that
from happening for eweapons.

Normally that wouldn't matter, but if the weapon was reflected it would
then try to run the eweapon script on the newly created lweapon.

And normally, _that_ would maybe mostly work, but another bug actually
caused a crash to happen instead.

So now eweapons have their script cleared when it ends, and the crash is
avoided.