refactor(zc)!: clean up optimizer/jit management
Adds a "zasm pipeline" system to manage when the optimizer or jit runs.

Also implements running the optimizer in the background as the game
runs, which mostly works but I've decided to remove it in the next
commit because it greatly complicated the code, and I think the
optimizer runs fast enough to just run on quest load.