feat(zscript): support negative indices for internal arrays
Some internal arrays would bound the index to force it to be valid,
while others would instead error and give a default value. Now, indices
are never bound, and instead if the index is negative it will "wrap
around" to the end of the array, just like all other arrays do.

Previous behavior is available behind a compat QR.