fix(zc): only force grab one held-up item at a time
If the player collects multiple items at or around the same time, and
those items run the "hold up" animation, only the first item would show
the holding animation while the others would silently be collected.

Now, the items are collected and held up one at a time. Note that this
only applies to "force grabbed" items (ex: settable in scripts via
`itemsprite::ForceGrab`, or as the result of some triggers).